Convert HtmlAttachmentPusher to a worker
We want to be able to call this logic asynchronously from Edition
services, so it needs to be converted to a Worker.
